requirements.txt
azure-kusto-data
azure-kusto-ingest
pylint
black
pandas
ipython
streamlit
numpy